Run ANALYZE in MmsSmsProvider.onOpen()
Up-to-date statistics make for more useful indices.

When adding an index, it gets stats based on the current data. So an
index added on a well-populated table worked well, but an index added
before any data would work poorly (the query planner wouldn't use it).

Let's update the stats periodically (on database open) with ANALYZE.

Change-Id: I556ce2bbb097435b7d5bec6a3c2a2b9cede979fa